+++ variables.tf
@@ -22,5 +22,4 @@ variable "allow_overwrite" {
 
 variable "tags" {
   type    = map(string)
-  default = {}
 }